java/src/main/java/com/yourtion/leetcode/daily/m10/d27/Solution.java
package com.yourtion.leetcode.daily.m10.d27;

import com.yourtion.leetcode.utils.TreeNode;

import java.util.ArrayList;
import java.util.List;

/**
 * 144. 二叉树的前序遍历
 *
 * @author Yourtion
 * @link https://leetcode-cn.com/problems/binary-tree-preorder-traversal/
 */
public class Solution {
    public List<Integer> preorderTraversal(TreeNode root) {
        List<Integer> res = new ArrayList<>();
        preorder(root, res);
        return res;
    }

    private void preorder(TreeNode root, List<Integer> res) {
        if (root == null) {
            return;
        }
        res.add(root.val);
        preorder(root.left, res);
        preorder(root.right, res);
    }
}
